Move class Cluster into cluster.py file
